Added PuppetCrashDialogHangler. Dialog will be opened from main thread.
