+++ Assets/Puppet/PuppetCrashDialogHandler.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 16e78ec24bddc624e9f7d7969e9b48d0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
