+++ Assets/Puppet/PuppetScene.unity
@@ -8930,7 +8930,7 @@ MonoBehaviour:
   m_HandleRect: {fileID: 1111163099}
   m_Direction: 2
   m_Value: 1
-  m_Size: 0.45983353
+  m_Size: 0.45805094
   m_NumberOfSteps: 0
   m_OnValueChanged:
     m_PersistentCalls:
@@ -21231,6 +21231,7 @@ GameObject:
   - component: {fileID: 1653541464}
   - component: {fileID: 1653541465}
   - component: {fileID: 1653541466}
+  - component: {fileID: 1653541463}
   m_Layer: 0
   m_Name: App Center
   m_TagString: Untagged
@@ -21238,6 +21239,18 @@ GameObject:
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
+--- !u!114 &1653541463
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 1653541462}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 16e78ec24bddc624e9f7d7969e9b48d0, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  ConfirmationDialog: {fileID: 1399467467}
 --- !u!114 &1653541464
 MonoBehaviour:
   m_ObjectHideFlags: 0
